Implement `stop` UCI command
The `stop` command should not only stop the current search, but also
send a `bestmove` message. So, the implementation runs a search
on a separate thread and moves the clone of `Arc<RwLock>` into it.
The `Arc` reference allows the search to handle exiting by reading
the shared mutable lock every 4096 search nodes in case it has changed
in the message input loop.