Simplify FEN error handling
Remove unnecessary custom error types in favor of using
`Box<dyn std::io::Error>` to represent FEN parsing errors.

After some consideration, it seems that the complexity of implementing
a proper error handling system is not justified.

The chess engine is not a library and is not intended to be used by
other programs. The only program that uses the chess engine is the GUI
via the UCI protocol.

While the GUI should ideally provide valid input to the chess engine,
it is important for the engine to handle invalid commands or similar
scenarios. Panicking upon receiving invalid input is not desirable.
Instead, the engine should return an appropriate error message and
ignore the invalid input.