Introduce basic prior-counter-move (pcm) history updates
The idea is that a move that make us try many moves and eventually make us fail low is probably good so boost its stats with hindsight.

the reasoning of a scaled bonus vs the normal update bonus can probably be related to the fact that the current depth of this ply is already decremented compared to the previous depth of the prior-counter-move it gets normally updated with.

Elo   | 3.87 +- 2.80 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.92 (-2.25, 2.89) [0.00, 4.00]
Games | N: 16414 W: 3957 L: 3774 D: 8683
Penta | [68, 1888, 4134, 2027, 90]
https://rickdiculous.pythonanywhere.com/test/3773/

bench: 5099312