Introduce static evaluation history (#233)
The idea is to update the history of a move based on how the static evaluation
changes after making the move. If it improves the evaluation, it's more likely to
be a good move, so the history is increased. The opposite also applies.

Elo   | 3.64 +- 2.66 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.95 (-2.25, 2.89) [0.00, 4.00]
Games | N: 18126 W: 4339 L: 4149 D: 9638
Penta | [67, 2136, 4482, 2296, 82]

Bench: 4054662