Fix bug in MVV scoring (#234)
Previously, a hack used the pawn value as the captured piece when the target
square was empty - this was meant to handle en passant, where the target square
is empty, but it's a pawn capture.
However, this also affects non-capture queen promotions, where the target square
is actually empty and should be treated as such.

Elo   | 1.07 +- 2.75 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.93 (-2.25, 2.89) [-5.00, 0.00]
Games | N: 15214 W: 3552 L: 3505 D: 8157
Penta | [51, 1637, 4209, 1634, 76]

Bench: 4174377